pyproject.toml
[tool.poetry]
name = "ylab-university-python-intensive-project"
version = "0.1.0"
description = ""
authors = ["Sergey Martynov <acemore@mail.ru>"]
readme = "README.md"
packages = [
  {include = "menu_app"},
  {include = "menu_app_tests"}
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
